Change the svelte page to use +page.server.ts to load data instead of an api route
Change the Svelte page to use `+page.server.ts` for data loading instead of an API route.

* **Add `src/routes/+page.server.ts`**:
  - Create a new file to handle server-side data fetching.
  - Define a `load` function to fetch user and friend profiles.
  - Fetch user profile using `fetchUser` function.
  - Fetch friend profiles using `fetchUser` function and return the data.

* **Modify `src/routes/+page.svelte`**:
  - Remove the `onMount` function and related data fetching logic.
  - Import `page` from `$app/stores`.
  - Use the data provided by `src/routes/+page.server.ts` to populate `profile` and `friendProfiles`.
  - Update the template to use the data from `page.data`.

* **Delete `src/routes/api/avatar/+server.ts`**:
  - Remove the file as it is no longer needed.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/link-discord/linkdiscord.xyz?shareId=XXXX-XXXX-XXXX-XXXX).
